First steps towards adding pre-generated source that the shell/makefile scripts typically take care of, but whose tools don't always work under every version of Windows.
This is by no means complete. It still needs to be integrated with the Visual Studio projects and any other auto-generated files need to be identified and added.

This is, unfortunately, a necessary evil for now until we can get an updated version of bison, flex, and the other tools that will work under multiple Windows versions and platforms.

git-svn-id: http://ossbuild.googlecode.com/svn/trunk@103 72a73a14-1cf0-11de-a027-6f5b8e30e268