Fixed small dependency issue (gstvideo depends on gstnet).
Fixed the installs according to the discussion (should be using bin/, lib/, shared/, etc. folders). Also added custom action to the installer to change the system path + add GST_PLUGIN_PATH env var.

Also fixed generate.bat. A file from Common/ was removed that was necessary for it to run correctly. That file has been re-introduced from the repo. It also doesn't require python to be installed. If python isn't present, it will just issue a warning and continue. Also fixed it to look for python in the system path in addition to C:\Python25\. Updated bison/flex scripts to not emit the line number.

Recompiled everything, ran the installers (LGPL and GPL) and everything worked! Holy cow! (c:

git-svn-id: http://ossbuild.googlecode.com/svn/trunk@302 72a73a14-1cf0-11de-a027-6f5b8e30e268