Fixes soup and png crashes.
Soup was crashing b/c it was unable to reach across the dll boundary and pull in a runtime-loaded "GET" string (SOUP_METHOD_GET).

Libpng was crashing b/c we were using the .def file for what would be a MSVC-compiled lib. The MinGW version was different. Using pexports on the MinGW-compiled dll and then creating a lib from that worked.

I didn't include the png12.lib b/c I'm still working on the new dependencies builds.

git-svn-id: http://ossbuild.googlecode.com/svn/trunk@330 72a73a14-1cf0-11de-a027-6f5b8e30e268